Up to date with develop

+++ app/src/main/java/ru/practicum/android/diploma/feature/filter/presentation/FilterRegionViewmodel.kt
@@ -67,7 +67,7 @@ class FilterRegionViewmodel(private val filterRegionsInteractor: FilterRegionsIn
         val selectedCountry = filterRegionsInteractor.getCountry()
 
         return if (selectedCountry != null) {
-            allRegions.filter { region -> region.parentId == selectedCountry.id.toInt() }
+            allRegions.filter { region -> region.parentId == selectedCountry.id }
         } else {
             allRegions
         }
